sort updates by message id, use first date in RSS
When a message is edited, it floated up in the Recent Updates column on
the front page, and its published date was updated. While somewhat
useful for humans, this feature caused many RSS readers to display
duplicate entries for new versions of the same message.

Now, edited messages will no longer float up. The published date will be
updated on the front page, but the RSS feed will instead use the first
date when the message was published.